requirements.txt
numpy
pillow
streamlit
tensorflow-cpu
tensorflow-hub
lime
skimage